// packages/core/src/components/special/DestructuredEditor/__generated__/getContent.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getContent
// ====================================================

export interface getContent_getContent {
  title: string | null;
  author: string | null;
  content: string;
}

export interface getContent {
  getContent: getContent_getContent;
}

export interface getContentVariables {
  contentId: string;
}